End of video 4
There is a problem with the path to load images and the fact that nor
the assets folder or the image file even if it is located within the src
code are not copied to the build directory then not loaded.

The player png file also needs to be improved, probably resized to at
least 64x64.